pyproject.toml
[project]
name = "nimble-research-harness"
version = "0.1.0"
description = "A production-grade dynamic web research agent platform powered by Nimble"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Urik" }]
dependencies = [
    "anthropic>=0.52.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "structlog>=24.1.0",
    "typer[all]>=0.12.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
api = ["fastapi>=0.111.0", "uvicorn[standard]>=0.30.0"]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0", "pytest-cov>=5.0.0", "ruff>=0.4.0"]

[project.scripts]
nrh = "nimble_research_harness.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nimble_research_harness"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
